+++ test/hook_widget_test.dart
@@ -10,6 +10,7 @@ void main() {
   final dispose = Func0<void>();
   final initHook = Func0<void>();
   final didUpdateHook = Func1<HookTest, void>();
+  final didBuild = Func0<void>();
   final reassemble = Func0<void>();
   final builder = Func1<BuildContext, Widget>();
 
@@ -19,11 +20,21 @@ void main() {
         didUpdateHook: didUpdateHook.call,
         reassemble: reassemble.call,
         initHook: initHook.call,
+        didBuild: didBuild,
       );
+
+  void verifyNoMoreHookInteration() {
+    verifyNoMoreInteractions(build);
+    verifyNoMoreInteractions(didBuild);
+    verifyNoMoreInteractions(dispose);
+    verifyNoMoreInteractions(initHook);
+    verifyNoMoreInteractions(didUpdateHook);
+  }
 
   tearDown(() {
     reset(builder);
     reset(build);
+    reset(didBuild);
     reset(dispose);
     reset(initHook);
     reset(didUpdateHook);
@@ -70,6 +81,7 @@ void main() {
         didUpdateHook: didUpdateHook.call,
         initHook: initHook.call,
         keys: keys,
+        didBuild: didBuild,
         createStateFn: createState.call,
       ));
       return Container();
@@ -82,24 +94,18 @@ void main() {
       createState.call(),
       initHook.call(),
       build.call(context),
+      didBuild.call(),
     ]);
-    verifyNoMoreInteractions(createState);
-    verifyNoMoreInteractions(initHook);
-    verifyNoMoreInteractions(build);
-    verifyNoMoreInteractions(didUpdateHook);
-    verifyNoMoreInteractions(dispose);
+    verifyNoMoreHookInteration();
 
     await tester.pumpWidget(HookBuilder(builder: builder.call));
 
     verifyInOrder([
       didUpdateHook.call(any),
       build.call(context),
+      didBuild.call(),
     ]);
-    verifyNoMoreInteractions(createState);
-    verifyNoMoreInteractions(initHook);
-    verifyNoMoreInteractions(build);
-    verifyNoMoreInteractions(didUpdateHook);
-    verifyNoMoreInteractions(dispose);
+    verifyNoMoreHookInteration();
 
     // from null to array
     keys = <dynamic>[];
@@ -109,13 +115,10 @@ void main() {
       dispose.call(),
       createState.call(),
       initHook.call(),
-      build.call(context)
+      build.call(context),
+      didBuild.call(),
     ]);
-    verifyNoMoreInteractions(createState);
-    verifyNoMoreInteractions(initHook);
-    verifyNoMoreInteractions(build);
-    verifyNoMoreInteractions(didUpdateHook);
-    verifyNoMoreInteractions(dispose);
+    verifyNoMoreHookInteration();
 
     // array immutable
     keys.add(42);
@@ -125,12 +128,9 @@ void main() {
     verifyInOrder([
       didUpdateHook.call(any),
       build.call(context),
+      didBuild.call(),
     ]);
-    verifyNoMoreInteractions(createState);
-    verifyNoMoreInteractions(initHook);
-    verifyNoMoreInteractions(build);
-    verifyNoMoreInteractions(didUpdateHook);
-    verifyNoMoreInteractions(dispose);
+    verifyNoMoreHookInteration();
 
     // new array but content equal
     keys = <dynamic>[42];
@@ -140,12 +140,9 @@ void main() {
     verifyInOrder([
       didUpdateHook.call(any),
       build.call(context),
+      didBuild.call(),
     ]);
-    verifyNoMoreInteractions(createState);
-    verifyNoMoreInteractions(initHook);
-    verifyNoMoreInteractions(build);
-    verifyNoMoreInteractions(didUpdateHook);
-    verifyNoMoreInteractions(dispose);
+    verifyNoMoreHookInteration();
 
     // new array new content
     keys = <dynamic>[44];
@@ -156,13 +153,10 @@ void main() {
       dispose.call(),
       createState.call(),
       initHook.call(),
-      build.call(context)
+      build.call(context),
+      didBuild.call()
     ]);
-    verifyNoMoreInteractions(createState);
-    verifyNoMoreInteractions(initHook);
-    verifyNoMoreInteractions(build);
-    verifyNoMoreInteractions(didUpdateHook);
-    verifyNoMoreInteractions(dispose);
+    verifyNoMoreHookInteration();
   });
 
   testWidgets('hook & setState', (tester) async {
@@ -189,6 +183,92 @@ void main() {
     expect(hookContext.dirty, true);
   });
 
+  testWidgets(
+      'didBuild when build crash called after FlutterError.onError report',
+      (tester) async {
+    final onError = FlutterError.onError;
+    FlutterError.onError = Func1<FlutterErrorDetails, void>();
+    final errorBuilder = ErrorWidget.builder;
+    ErrorWidget.builder = Func1<FlutterErrorDetails, Widget>();
+    when(ErrorWidget.builder(any)).thenReturn(Container());
+    try {
+      when(build.call(any)).thenThrow(42);
+      when(builder.call(any)).thenAnswer((invocation) {
+        Hook.use(createHook());
+        return Container();
+      });
+
+      await tester.pumpWidget(HookBuilder(
+        builder: builder.call,
+      ));
+      tester.takeException();
+
+      verifyInOrder([
+        build.call(any),
+        FlutterError.onError(any),
+        ErrorWidget.builder(any),
+        didBuild(),
+      ]);
+    } finally {
+      FlutterError.onError = onError;
+      ErrorWidget.builder = errorBuilder;
+    }
+  });
+
+  testWidgets('didBuild called even if build crashed', (tester) async {
+    when(build.call(any)).thenThrow(42);
+    when(builder.call(any)).thenAnswer((invocation) {
+      Hook.use(createHook());
+      return Container();
+    });
+
+    await tester.pumpWidget(HookBuilder(
+      builder: builder.call,
+    ));
+    expect(tester.takeException(), 42);
+
+    verify(didBuild.call()).called(1);
+  });
+  testWidgets('all didBuild called even if one crashes', (tester) async {
+    final didBuild2 = Func0<void>();
+
+    when(didBuild.call()).thenThrow(42);
+    when(builder.call(any)).thenAnswer((invocation) {
+      Hook.use(createHook());
+      Hook.use(HookTest<int>(didBuild: didBuild2));
+      return Container();
+    });
+
+    await expectPump(
+      () => tester.pumpWidget(HookBuilder(
+            builder: builder.call,
+          )),
+      throwsA(42),
+    );
+
+    verifyInOrder([
+      didBuild2.call(),
+      didBuild.call(),
+    ]);
+  });
+
+  testWidgets('calls didBuild before building children', (tester) async {
+    final buildChild = Func1<BuildContext, Widget>();
+    when(buildChild.call(any)).thenReturn(Container());
+
+    await tester.pumpWidget(HookBuilder(
+      builder: (context) {
+        Hook.use(createHook());
+        return Builder(builder: buildChild);
+      },
+    ));
+
+    verifyInOrder([
+      didBuild(),
+      buildChild.call(any),
+    ]);
+  });
+
   testWidgets('life-cycles in order', (tester) async {
     int result;
     HookTest<int> hook;
@@ -209,9 +289,9 @@ void main() {
     verifyInOrder([
       initHook.call(),
       build.call(context),
+      didBuild.call(),
     ]);
-    verifyZeroInteractions(didUpdateHook);
-    verifyZeroInteractions(dispose);
+    verifyNoMoreHookInteration();
 
     when(build.call(context)).thenReturn(24);
     var previousHook = hook;
@@ -221,28 +301,19 @@ void main() {
     ));
 
     expect(result, 24);
-    verifyInOrder([
-      didUpdateHook.call(previousHook),
-      build.call(any),
-    ]);
-    verifyNoMoreInteractions(initHook);
-    verifyZeroInteractions(dispose);
+    verifyInOrder(
+        [didUpdateHook.call(previousHook), build.call(any), didBuild.call()]);
+    verifyNoMoreHookInteration();
 
     previousHook = hook;
     await tester.pump();
 
-    verifyNoMoreInteractions(initHook);
-    verifyNoMoreInteractions(didUpdateHook);
-    verifyNoMoreInteractions(build);
-    verifyZeroInteractions(dispose);
+    verifyNoMoreHookInteration();
 
     await tester.pumpWidget(const SizedBox());
 
-    verifyNoMoreInteractions(initHook);
-    verifyNoMoreInteractions(didUpdateHook);
-    verifyNoMoreInteractions(build);
-    verify(dispose.call());
-    verifyNoMoreInteractions(dispose);
+    verify(dispose.call()).called(1);
+    verifyNoMoreHookInteration();
   });
 
   testWidgets('dispose all called even on failed', (tester) async {