+++ test/use_stream_controller_test.dart
@@ -17,7 +17,7 @@ void main() {
 
       final previous = controller;
       await tester.pumpWidget(HookBuilder(builder: (context) {
-        controller = useStreamController(keys: <dynamic>[]);
+        controller = useStreamController(keys: []);
         return Container();
       }));