+++ lib/src/framework.dart
@@ -325,7 +325,8 @@ This may happen if the call to `Hook.use` is made under some condition.
             exception: exception,
             stack: stack,
             library: 'hooks library',
-            context: 'while calling `didBuild` on ${hook.runtimeType}',
+            context: ErrorDescription(
+                'while calling `didBuild` on ${hook.runtimeType}'),
           ));
         }
       }
@@ -345,7 +346,7 @@ This may happen if the call to `Hook.use` is made under some condition.
             exception: exception,
             stack: stack,
             library: 'hooks library',
-            context: 'while disposing ${hook.runtimeType}',
+            context: ErrorDescription('while disposing ${hook.runtimeType}'),
           ));
         }
       }